fix bug

+++ Visualize_attentions.py
@@ -46,7 +46,7 @@ def visualize(im,logits, att_mat):
     mask = v[0, 1:].reshape(grid_size, grid_size).detach().numpy()
     pdb.set_trace()
     mask = cv2.resize(mask / mask.max(), (im.shape[1], im.shape[2]))[..., np.newaxis]
-    result = (mask * im.numpy().transpose(1,2,0)).astype("uint8")
+    result = (mask * im.numpy().transpose(1,2,0)).astype("float64")
     np.save("image_2.npy", result)
     np.save("image_original_2.npy", im.numpy().transpose(1,2,0))
